RavenDB-15885 added test
