+++ test/EmbeddedTests/BasicTests.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Diagnostics;
-using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using Raven.Client.Documents.Conventions;
@@ -121,59 +120,6 @@ namespace EmbeddedTests
                 Assert.NotEqual(pid2, pid3);
             }
         }
-
-        private (string ServerDirectory, string DataDirectory) CopyServer()
-        {
-            var baseDirectory = NewDataPath();
-            var serverDirectory = Path.Combine(baseDirectory, "RavenDBServer");
-            var dataDirectory = Path.Combine(baseDirectory, "RavenDB");
-
-            if (Directory.Exists(serverDirectory) == false)
-                Directory.CreateDirectory(serverDirectory);
-
-            if (Directory.Exists(dataDirectory) == false)
-                Directory.CreateDirectory(dataDirectory);
-
-#if DEBUG
-            var runtimeConfigPath = @"../../../../../src/Raven.Server/bin/x64/Debug/netcoreapp3.1/Raven.Server.runtimeconfig.json";
-            if (File.Exists(runtimeConfigPath) == false) // this can happen when running directly from CLI e.g. dotnet xunit
-                runtimeConfigPath = @"../../../../../src/Raven.Server/bin/Debug/netcoreapp3.1/Raven.Server.runtimeconfig.json";
-#else
-                var runtimeConfigPath = @"../../../../../src/Raven.Server/bin/x64/Release/netcoreapp3.1/Raven.Server.runtimeconfig.json";
-                if (File.Exists(runtimeConfigPath) == false) // this can happen when running directly from CLI e.g. dotnet xunit
-                    runtimeConfigPath = @"../../../../../src/Raven.Server/bin/Release/netcoreapp3.1/Raven.Server.runtimeconfig.json";
-#endif
-
-            var runtimeConfigFileInfo = new FileInfo(runtimeConfigPath);
-            if (runtimeConfigFileInfo.Exists == false)
-                throw new FileNotFoundException("Could not find runtime config", runtimeConfigPath);
-
-            File.Copy(runtimeConfigPath, Path.Combine(serverDirectory, runtimeConfigFileInfo.Name), true);
-
-            foreach (var extension in new[] { "*.dll", "*.so", "*.dylib", "*.deps.json" })
-            {
-                foreach (var file in Directory.GetFiles(runtimeConfigFileInfo.DirectoryName, extension))
-                {
-                    var fileInfo = new FileInfo(file);
-                    File.Copy(file, Path.Combine(serverDirectory, fileInfo.Name), true);
-                }
-            }
-
-            var runtimesSource = Path.Combine(runtimeConfigFileInfo.DirectoryName, "runtimes");
-            var runtimesDestination = Path.Combine(serverDirectory, "runtimes");
-
-            foreach (string dirPath in Directory.GetDirectories(runtimesSource, "*",
-                SearchOption.AllDirectories))
-                Directory.CreateDirectory(dirPath.Replace(runtimesSource, runtimesDestination));
-
-            foreach (string newPath in Directory.GetFiles(runtimesSource, "*.*",
-                SearchOption.AllDirectories))
-                File.Copy(newPath, newPath.Replace(runtimesSource, runtimesDestination), true);
-
-            return (serverDirectory, dataDirectory);
-        }
-
-
 
         private class Person
         {
